Start gunicorn directly in Docker image
Using a bash shell to run a script that simply runs gunicorn means that
the process with PID 1 in the container is now bash, which doesn't have
a SIGTERM handler installed by default; hence at pod shutdown the
container doesn't finish immediately, and we have to wait for a 30
second timeout for it to be SIGKILL'ed. gunicorn on the other hand does
have a SIGTERM handler, and therefore is better suited to be the main
process running on the container.

This was found while working on YAN-951.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>